backend/requirements.txt
azure-functions==1.18.0
azure-identity==1.15.0
azure-ai-inference==1.0.0b1
azure-search-documents==11.4.0
msal==1.26.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
azure-monitor-opentelemetry==1.2.0
azure-monitor-opentelemetry-exporter==1.0.0b21
openai==1.12.0
pydantic==2.6.0
